+++ src/fides/cli/commands/core.py
@@ -4,7 +4,8 @@ from typing import Optional
 import rich_click as click
 
 from fides.cli.options import dry_flag, manifests_dir_argument, verbose_flag
-from fides.cli.utils import pretty_echo, print_divider, with_analytics
+from fides.cli.utils import with_analytics
+from fides.common.utils import pretty_echo, print_divider
 from fides.core import audit as _audit
 from fides.core import evaluate as _evaluate
 from fides.core import parse as _parse
