+++ src/fides/core/api_helpers.py
@@ -10,7 +10,7 @@ from fideslang.validation import FidesKey
 from requests import Response
 
 from fides.core import api
-from fides.core.utils import check_response_auth
+from fides.common.utils import check_response_auth
 
 
 def get_server_resources(
