+++ src/fides/core/evaluate.py
@@ -21,7 +21,7 @@ from fideslang.utils import get_resource_by_fides_key
 from fideslang.validation import FidesKey
 from pydantic import AnyHttpUrl
 
-from fides.cli.utils import handle_cli_response, pretty_echo
+from fides.common.utils import handle_cli_response, pretty_echo
 from fides.core import api
 from fides.core.api_helpers import get_server_resource, get_server_resources
 from fides.core.parse import parse
