+++ src/fides/core/pull.py
@@ -5,7 +5,7 @@ import yaml
 from fideslang import model_list
 from fideslang.manifests import load_yaml_into_dict
 
-from fides.cli.utils import print_divider
+from fides.common.utils import print_divider
 from fides.core.api_helpers import get_server_resource, list_server_resources
 from fides.core.utils import echo_green, get_manifest_list
