+++ src/fides/core/user.py
@@ -6,7 +6,7 @@ import requests
 from fideslang.validation import FidesKey
 
 from fides.api.ops.cryptography.cryptographic_util import str_to_b64_str
-from fides.cli.utils import handle_cli_response
+from fides.common.utils import handle_cli_response
 from fides.core.config import CONFIG
 from fides.core.utils import (
     Credentials,
